Update: refactor project structure to use 'app' namespace and enhance dashboard history visualization
