+++ app/api/v1/router.py
@@ -1,7 +1,7 @@
 # weather_app/api/v1/router.py
 
 from fastapi import APIRouter
-from weather_app.api.v1.endpoints import weather
+from app.api.v1.endpoints import weather
 
 api_router = APIRouter()
