+++ app/templates/history.html
@@ -1,0 +1,59 @@
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>Historia pomiarów</title>
+    <style>
+        body { font-family: Arial; margin: 40px; }
+        .tabs button {
+            padding: 10px;
+            cursor: pointer;
+            background-color: lightgray;
+            border: none;
+            outline: none;
+        }
+        .tabs button.active {
+            background-color: #ddd;
+            font-weight: bold;
+        }
+        .tab-content {
+            display: none;
+            margin-top: 20px;
+        }
+        .tab-content.active {
+            display: block;
+        }
+        img { max-width: 100%; }
+    </style>
+  </head>
+<body>
+ <h1>Historia pomiarów (7 dni)</h1>Add commentMore actions
+
+    <div class="tabs">
+        <button onclick="showTab('temp')" class="active">Temperatura i ciśnienie</button>
+        <button onclick="showTab('rain')">Opady</button>
+        <button onclick="showTab('wind')">Wiatr</button>
+    </div>
+
+    <div id="temp" class="tab-content active">
+        <h3>Temperatura i ciśnienie</h3>
+        <img src="data:image/png;base64,{{ plot_temp_pressure }}">
+    </div>
+
+    <div id="rain" class="tab-content">
+        <h3>Opady</h3>
+        <img src="data:image/png;base64,{{ plot_rain }}">
+    </div>
+<div id="wind" class="tab-content">
+        <h3>Wiatr</h3>
+        <img src="data:image/png;base64,{{ plot_wind }}">
+    </div>
+<script>Add commentMore actions
+        function showTab(id) {
+            document.querySelectorAll('.tab-content').forEach(div => div.classList.remove('active'));
+            document.querySelectorAll('.tabs button').forEach(btn => btn.classList.remove('active'));
+            document.getElementById(id).classList.add('active');
+            event.target.classList.add('active');
+        }
+    </script>
+</body>
+</html>
