--- weather_app/templates/history.html
@@ -1,66 +1,0 @@
-<!-- weather_app/templates/history.html -->
-
-<!DOCTYPE html>
-<html lang="pl">
-<head>
-  <meta charset="UTF-8">
-  <title>Historia pomiarów</title>
-  <style>
-    body {
-      font-family: Arial, sans-serif;
-      padding: 20px;
-    }
-    img {
-      max-width: 100%;
-    }
-    table {
-      width: 100%;
-      border-collapse: collapse;
-      margin-top: 20px;
-    }
-    th, td {
-      padding: 8px;
-      border-bottom: 1px solid #ccc;
-      text-align: center;
-    }
-    th {
-      background-color: #f2f2f2;
-    }
-  </style>
-</head>
-<body>
-  <h1>Historia pomiarów z ostatnich 7 dni</h1>
-
-  {% if plot_image %}
-    <img src="data:image/png;base64,{{ plot_image }}" alt="Wykres historii">
-  {% else %}
-    <p>Brak danych do wyświetlenia wykresu.</p>
-  {% endif %}
-
-  <h2>Szczegóły pomiarów</h2>
-  <table>
-    <thead>
-      <tr>
-        <th>Data i czas</th>
-        <th>Temperatura (°C)</th>
-        <th>Wilgotność (%)</th>
-        <th>Ciśnienie (hPa)</th>
-        <th>Opady (mm)</th>
-        <th>Wiatr (km/h)</th>
-      </tr>
-    </thead>
-    <tbody>
-      {% for record in history %}
-      <tr>
-        <td>{{ record.timestamp.strftime("%Y-%m-%d %H:%M:%S") }}</td>
-        <td>{{ record.temperature }}</td>
-        <td>{{ record.humidity }}</td>
-        <td>{{ record.pressure }}</td>
-        <td>{{ record.rainfall }}</td>
-        <td>{{ record.wind_speed }}</td>
-      </tr>
-      {% endfor %}
-    </tbody>
-  </table>
-</body>
-</html>
